refactor: RandomPick Modal 수정
